requirements.txt
prometheus-client==0.11.0
PyYAML==6.0
speedtest-cli==2.1.3
